+++ sdk/clientcore/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.8 (Unreleased)
+## 1.0.0-beta.9 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,17 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0-beta.8 (2025-04-03)
+
+### Breaking Changes
+- Redesign RequestOptions - merge with context [#44535](https://github.com/Azure/azure-sdk-for-java/pull/44535)
+- API updates to refactor [#44655](https://github.com/Azure/azure-sdk-for-java/pull/44655)
+- Remove ResponseBodyMode [#44635](https://github.com/Azure/azure-sdk-for-java/pull/44635)
+- Remove RestProxy code [#44660](https://github.com/Azure/azure-sdk-for-java/pull/44660)
+
+### Bugs Fixed
+- Fix vNext Bearer Token Policy [#44803](https://github.com/Azure/azure-sdk-for-java/pull/44803/)
 
 ## 1.0.0-beta.7 (2025-03-12)