+++ sdk/keyvault/azure-security-keyvault-administration/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.7.0-beta.1 (2025-03-24)
+## 4.7.0-beta.1 (2025-04-09)
 
 ### Features Added
 - Added support for pre-backup and pre-restore operations in `KeyVaultBackupClient` and `KeyVaultBackupAsyncClient`. These are long-running operations that check whether it is possible to perform a full key backup or a full key restore from a backup. ([#39878](https://github.com/Azure/azure-sdk-for-java/pull/39878))
@@ -12,6 +12,16 @@
 - Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
 - Upgraded `azure-json` from `1.4.0` to version `1.5.0`.
 - Upgraded `azure-core-http-netty` from `1.15.10` to version `1.15.11`.
+
+## 4.6.4 (2025-03-24)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-json` from `1.4.0` to version `1.5.0`.
+- Upgraded `azure-core-http-netty` from `1.15.10` to version `1.15.11`.
+- Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
 
 ## 4.6.3 (2025-03-04)