+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.10.0-beta.1 (2025-03-24)
+## 4.10.0-beta.1 (2025-04-09)
 
 ### Features Added
 - Added new key wrapping algorithms: `CKM_AES_KEY_WRAP` and `CKM_AES_KEY_WRAP_PAD`. ([#44788](https://github.com/Azure/azure-sdk-for-java/pull/44788)
@@ -14,6 +14,16 @@
 - Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
 - Upgraded `azure-json` from `1.4.0` to version `1.5.0`.
 - Upgraded `azure-core-http-netty` from `1.15.10` to version `1.15.11`.
+
+## 4.9.4 (2025-03-24)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-json` from `1.4.0` to version `1.5.0`.
+- Upgraded `azure-core-http-netty` from `1.15.10` to version `1.15.11`.
+- Upgraded `azure-core` from `1.55.2` to version `1.55.3`.
 
 ## 4.9.3 (2025-03-04)
 
@@ -68,7 +78,6 @@
 
 - Upgraded `azure-core` from `1.50.0` to version `1.51.0`.
 - Upgraded `azure-core-http-netty` from `1.15.2` to version `1.15.3`.
-
 
 ## 4.9.0-beta.1 (2024-07-29)