+++ sdk/resourcemanager/azure-resourcemanager-keyvault/CHANGELOG.md
@@ -1,10 +1,22 @@
 # Release History
 
-## 2.49.0-beta.1 (Unreleased)
+## 2.50.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 2.49.0 (2025-03-28)
+
+### Other Changes
 
 #### Dependency Updates
 
-- Updated `api-version` to `2024-11-01`.
+- Updated core dependency from resources.
 
 ## 2.48.0 (2025-02-28)